+++ crates/alkahest/src/gui/configuration.rs
@@ -2,7 +2,7 @@ use alkahest_renderer::{
     camera::{Camera, CameraProjection},
     ecs::tags::{NodeFilter, NodeFilterSet},
     icons::{ICON_CLIPBOARD, ICON_CURSOR_DEFAULT, ICON_EYE},
-    renderer::{RenderDebugView, RenderFeatureVisibility, RendererShared},
+    renderer::{RenderDebugView, RenderFeatureVisibility, RendererShared, ShadowQuality},
     util::text::StringExt,
 };
 use egui::{Context, RichText, Rounding, Widget};
@@ -101,7 +101,18 @@ impl GuiView for RenderSettingsPanel {
                 ui.collapsing(RichText::new("Graphics").heading(), |ui| {
                     ui.checkbox(&mut c.renderer.vsync, "VSync");
                     ui.checkbox(&mut c.renderer.matcap, "Matcap");
-                    ui.checkbox(&mut c.renderer.shadows, "Shadows");
+
+                    egui::ComboBox::from_label("Shadows")
+                        .selected_text(c.renderer.shadow_quality.to_string().split_pascalcase())
+                        .show_ui(ui, |ui| {
+                            for quality in ShadowQuality::iter() {
+                                ui.selectable_value(
+                                    &mut c.renderer.shadow_quality,
+                                    quality,
+                                    quality.to_string().split_pascalcase(),
+                                );
+                            }
+                        });
                     ui.checkbox(&mut c.renderer.ssao, "SSAO");
                     ui.collapsing("SSAO Settings", |ui| {
                         let renderer = resources.get::<RendererShared>();