+++ CHANGELOG.md
@@ -45,6 +45,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 - Added console command `window_resize` to resize the window to a specific size
 - Added console command `set_camera_from_cb12` to load view matrices from a raw view scope buffer
 - Added commands `lock_time` and `unlock_time` to allow fixing the game time to a specific value (eg. for more deterministic image comparisons)
+- Shadow quality option (replaces shadow checkbox)
 
 ### Changed